start fixing coordinate system for CMZ density
the correct coordinate system has the x-axis pointing towards the GC and the y-axis points to l=-90°.
For now the y-axis was not correctly orientated.

This fix only includes the diffuse medium. Also the contribution of HI and H2 are seperated.

NOTE: The molecular clouds can not be used with this commit.